ecs: Move EcsCompId to its own header

+++ libs/ecs/src/buffer.c
@@ -153,7 +153,7 @@ static EcsBufferCompData* ecs_buffer_compdata_add(
 EcsBuffer ecs_buffer_create(Allocator* alloc, const EcsDef* def) {
   return (EcsBuffer){
       .def      = def,
-      .masks    = dynarray_create(alloc, bits_to_bytes(ecs_def_comp_count(def)) + 1, 1, 256),
+      .masks    = dynarray_create(alloc, (u16)bits_to_bytes(ecs_def_comp_count(def)) + 1, 1, 256),
       .entities = dynarray_create_t(alloc, EcsBufferEntity, 256),
       .compDataAllocator =
           alloc_chunked_create(g_alloc_page, alloc_bump_create, ecs_buffer_compdata_chunk_size),
